requirements.txt
aiosqlite>=0.19.0
alembic>=1.18.4,<2.0.0
astral>=3.2
# Core production dependencies for Darkstar
# Database
fastapi>=0.109.0
httpx>=0.27.0
lightgbm>=4.0.0
nordpool>=0.2
open-meteo-solar-forecast
pandas>=2.0.0
pulp>=2.9.0
pydantic>=2.0.0
python-dateutil>=2.8.0
python-json-logger>=2.0.7
python-socketio>=5.11.0
pytz>=2023.3
pyyaml>=6.0
requests>=2.0.0
aiohttp>=3.9.0
ruamel.yaml>=0.17.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.48,<3.0.0
uvicorn[standard]>=0.27.0
websockets>=11.0
